some changes 2 in HW01

+++ src/main/java/ru/javawebinar/topjava/repository/UserMealRepository.java
@@ -11,7 +11,7 @@ public interface UserMealRepository {
 
     void delete(UserMeal userMeal);
 
-    UserMeal get(int id);
+    UserMeal get(long id);
 
     Collection<UserMeal> getAll ();
